requirements.txt
requests >= 2.28.1
pyOpenSSL >= 22.0.0
boto3 >= 1.24.55
botocore >= 1.27.55 
cryptography >= 37.0.4